scripts/images/requirements.txt
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0
python-slugify>=8.0

